프로그램 전체에서 값 검증을 수행하는 Validator 작성
